requirements-test.txt
coveralls
django-redis
django-cors-headers~=3.4
openwisp-utils[qa]~=0.6
redis
channels_redis
pytest-asyncio
pytest-django
